pp_scraper/requirements.txt


# Core scraping and HTTP
selenium>=4.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0

# Google Cloud BigQuery
google-cloud-bigquery>=3.0.0

# Optional: for automatic ChromeDriver management (recommended)
webdriver-manager>=3.8.0